refactor: move components files -> common
