+++ src/app/(auth-pages)/forgot-password/page.tsx
@@ -1,4 +1,4 @@
-import type { Message } from "@/components/form-message";
+import type { Message } from "@/components/common/form-message";
 import { ForgotPasswordForm } from "@/features/auth/components/forgot-password-form";
 import Image from "next/image";
