+++ src/app/(auth-pages)/sign-in/page.tsx
@@ -1,4 +1,4 @@
-import { FormMessage, type Message } from "@/components/form-message";
+import { FormMessage, type Message } from "@/components/common/form-message";
 import SignInForm from "@/features/auth/components/sign-in-form";
 import Image from "next/image";
 import Link from "next/link";
