+++ src/app/(auth-pages)/sign-up-email/page.tsx
@@ -1,4 +1,4 @@
-import type { Message } from "@/components/form-message";
+import type { Message } from "@/components/common/form-message";
 import EmailSignUpForm from "@/features/auth/components/email-sign-up-form";
 import Image from "next/image";
