+++ src/components/common/collapsible-info.tsx
