+++ src/components/common/env-var-warning.tsx
@@ -1,6 +1,6 @@
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
 import Link from "next/link";
-import { Badge } from "./ui/badge";
-import { Button } from "./ui/button";
 
 export function EnvVarWarning() {
   return (
