+++ src/components/common/notice-board-alert.tsx
