+++ src/features/auth/components/email-sign-up-form.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { emailSignUpActionWithState } from "@/app/actions";
-import { FormMessage, type Message } from "@/components/form-message";
-import { SubmitButton } from "@/components/submit-button";
+import { FormMessage, type Message } from "@/components/common/form-message";
+import { SubmitButton } from "@/components/common/submit-button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { passwordAlertlessSchema } from "@/lib/validation/auth";
